requirements.txt
# This file lists the dependencies of this extension.
# Install with a command like: pip install -r requirements.txt 
lxml>=4.4.0
requests>=2.5.3
ckanapi>=3.4
six>=1.12.0
